change module names
