+++ scripts/nmt/transformer.py
@@ -1,19 +1,3 @@
-"""
-Transformer
-=================================
-
-This example shows how to implement the Transformer model with Gluon NLP Toolkit.
-
-@inproceedings{vaswani2017attention,
-  title={Attention is all you need},
-  author={Vaswani, Ashish and Shazeer, Noam and Parmar, Niki and Uszkoreit, Jakob and Jones,
-          Llion and Gomez, Aidan N and Kaiser, Lukasz and Polosukhin, Illia},
-  booktitle={Advances in Neural Information Processing Systems},
-  pages={6000--6010},
-  year={2017}
-}
-"""
-
 # coding: utf-8
 
 # Licensed to the Apache Software Foundation (ASF) under one
@@ -32,517 +16,824 @@ This example shows how to implement the Transformer model with Gluon NLP Toolkit
 # KIND, either express or implied.  See the License for the
 # specific language governing permissions and limitations
 # under the License.
-# pylint:disable=redefined-outer-name,logging-format-interpolation
+"""Encoder and decoder usded in sequence-to-sequence learning."""
+__all__ = ['TransformerEncoder', 'TransformerDecoder', 'get_transformer_encoder_decoder']
 
-import argparse
-import time
-import random
-import os
-import io
-import logging
 import math
 import numpy as np
 import mxnet as mx
-from mxnet import gluon
-from mxnet.gluon.data import ArrayDataset, SimpleDataset
-from mxnet.gluon.data import DataLoader
-import gluonnlp.data.batchify as btf
-from gluonnlp.data import FixedBucketSampler, IWSLT2015, WMT2016BPE
-from gluonnlp.model import BeamSearchScorer
-from translation import NMTModel, BeamSearchTranslator
-from transformer_encoder_decoder import get_transformer_encoder_decoder
-from loss import SoftmaxCEMaskedLoss, LabelSmoothing
-from utils import logging_config
-from bleu import compute_bleu
-import _constants as _C
-
-np.random.seed(100)
-random.seed(100)
-mx.random.seed(10000)
-
-parser = argparse.ArgumentParser(description='Neural Machine Translation Example.'
-                                             'We train the Transformer Model')
-parser.add_argument('--dataset', type=str, default='WMT2016BPE', help='Dataset to use.')
-parser.add_argument('--src_lang', type=str, default='en', help='Source language')
-parser.add_argument('--tgt_lang', type=str, default='de', help='Target language')
-parser.add_argument('--epochs', type=int, default=10, help='upper epoch limit')
-parser.add_argument('--num_units', type=int, default=512, help='Dimension of the embedding '
-                                                               'vectors and states.')
-parser.add_argument('--hidden_size', type=int, default=2048,
-                    help='Dimension of the hidden state in position-wise feed-forward networks.')
-parser.add_argument('--dropout', type=float, default=0.1,
-                    help='dropout applied to layers (0 = no dropout)')
-parser.add_argument('--epsilon', type=float, default=0.1,
-                    help='epsilon parameter for label smoothing')
-parser.add_argument('--num_layers', type=int, default=6,
-                    help='number of layers in the encoder and decoder')
-parser.add_argument('--num_heads', type=int, default=8,
-                    help='number of heads in multi-head attention')
-parser.add_argument('--scaled', type=bool, default=True, help='whether to use scale in attention')
-parser.add_argument('--batch_size', type=int, default=1024,
-                    help='Batch size. Number of tokens in a minibatch')
-parser.add_argument('--beam_size', type=int, default=4, help='Beam size')
-parser.add_argument('--lp_alpha', type=float, default=0.6,
-                    help='Alpha used in calculating the length penalty')
-parser.add_argument('--lp_k', type=int, default=5, help='K used in calculating the length penalty')
-parser.add_argument('--test_batch_size', type=int, default=256, help='Test batch size')
-parser.add_argument('--num_buckets', type=int, default=10, help='Bucket number')
-parser.add_argument('--bucket_ratio', type=float, default=0.0, help='Ratio for increasing the '
-                                                                    'throughput of the bucketing')
-parser.add_argument('--src_max_len', type=int, default=-1, help='Maximum length of the source '
-                                                                'sentence, -1 means no clipping')
-parser.add_argument('--tgt_max_len', type=int, default=-1, help='Maximum length of the target '
-                                                                'sentence, -1 means no clipping')
-parser.add_argument('--optimizer', type=str, default='adam', help='optimization algorithm')
-parser.add_argument('--lr', type=float, default=1.0, help='Initial learning rate')
-parser.add_argument('--warmup_steps', type=float, default=4000,
-                    help='number of warmup steps used in NOAM\'s stepsize schedule')
-parser.add_argument('--num_accumulated', type=int, default=1,
-                    help='Number of steps to accumulate the gradients. '
-                         'This is useful to mimic large batch training with limited gpu memory')
-parser.add_argument('--magnitude', type=float, default=3.0,
-                    help='Magnitude of Xavier initialization')
-parser.add_argument('--average_checkpoint', type=bool, default=False,
-                    help='Whether to perform final testing based on '
-                         'the average of last few checkpoints')
-parser.add_argument('--num_averages', type=int, default=5,
-                    help='Perform final testing based on the '
-                         'average of last num_averages checkpoints. '
-                         'This is only used if average_checkpoint is True')
-parser.add_argument('--average_start', type=int, default=5,
-                    help='Perform average SGD on last average_start epochs')
-parser.add_argument('--log_interval', type=int, default=100, metavar='N',
-                    help='report interval')
-parser.add_argument('--save_dir', type=str, default='transformer_out',
-                    help='directory path to save the final model and training log')
-parser.add_argument('--gpus', type=str,
-                    help='list of gpus to run, e.g. 0 or 0,2,5. empty means using cpu.'
-                         '(using single gpu is suggested)')
-args = parser.parse_args()
-logging_config(args.save_dir)
-logging.info(args)
+from mxnet.gluon import nn
+from mxnet.gluon.block import HybridBlock
+from encoder_decoder import Seq2SeqEncoder, Seq2SeqDecoder, _get_attention_cell
 
 
-def cache_dataset(dataset, prefix):
-    """Cache the processed npy dataset  the dataset into a npz
+def _position_encoding_init(max_length, dim):
+    """ Init the sinusoid position encoding table """
+    position_enc = np.arange(max_length).reshape((-1, 1)) \
+                   / (np.power(10000, (2. / dim) * np.arange(dim).reshape((1, -1))))
+    # Apply the cosine to even columns and sin to odds.
+    position_enc[:, 0::2] = np.sin(position_enc[:, 0::2])  # dim 2i
+    position_enc[:, 1::2] = np.cos(position_enc[:, 1::2])  # dim 2i+1
+    return position_enc
+
+
+class PositionwiseFFN(HybridBlock):
+    """Structure of the Positionwise Feed-Forward Neural Network.
 
     Parameters
     ----------
-    dataset : SimpleDataset
-    file_path : str
+    units : int
+    hidden_size : int
+        number of units in the hidden layer of position-wise feed-forward networks
+    dropout : float
+    use_residual : bool
+    weight_initializer : str or Initializer
+        Initializer for the input weights matrix, used for the linear
+        transformation of the inputs.
+    bias_initializer : str or Initializer
+        Initializer for the bias vector.
+    activation : str, default 'relu'
+        Activation function
+    prefix : str, default 'rnn_'
+        Prefix for name of `Block`s
+        (and name of weight if params is `None`).
+    params : Parameter or None
+        Container for weight sharing between cells.
+        Created if `None`.
     """
-    if not os.path.exists(_C.CACHE_PATH):
-        os.makedirs(_C.CACHE_PATH)
-    src_data = np.array([ele[0] for ele in dataset])
-    tgt_data = np.array([ele[1] for ele in dataset])
-    np.savez(os.path.join(_C.CACHE_PATH, prefix + '.npz'), src_data=src_data, tgt_data=tgt_data)
+    def __init__(self, units=512, hidden_size=2048, dropout=0.0, use_residual=True,
+                 weight_initializer=None, bias_initializer='zeros', activation='relu',
+                 prefix=None, params=None):
+        super(PositionwiseFFN, self).__init__(prefix=prefix, params=params)
+        self._hidden_size = hidden_size
+        self._units = units
+        self._use_residual = use_residual
+        self.ffn_1 = nn.Dense(units=hidden_size, flatten=False,
+                              activation=activation,
+                              weight_initializer=weight_initializer,
+                              bias_initializer=bias_initializer,
+                              prefix='ffn_1_')
+        self.ffn_2 = nn.Dense(units=units, flatten=False,
+                              weight_initializer=weight_initializer,
+                              bias_initializer=bias_initializer,
+                              prefix='ffn_2_')
+        self.dropout_layer = nn.Dropout(dropout)
+        self.layer_norm = nn.LayerNorm()
+
+    def hybrid_forward(self, F, inputs):  # pylint: disable=unused-argument
+        """Position-wise encoding of the inputs.
+
+        Parameters
+        ----------
+        inputs : Symbol or NDArray
+            Input sequence. Shape (batch_size, length, C_in)
+
+        Returns
+        -------
+        outputs : Symbol or NDArray
+            Shape (batch_size, length, C_out)
+        """
+        outputs = self.ffn_1(inputs)
+        outputs = self.ffn_2(outputs)
+        outputs = self.dropout_layer(outputs)
+        if self._use_residual:
+            outputs = outputs + inputs
+        outputs = self.layer_norm(outputs)
+        return outputs
 
 
-def load_cached_dataset(prefix):
-    cached_file_path = os.path.join(_C.CACHE_PATH, prefix + '.npz')
-    if os.path.exists(cached_file_path):
-        print('Load cached data from {}'.format(cached_file_path))
-        dat = np.load(cached_file_path)
-        return ArrayDataset(np.array(dat['src_data']), np.array(dat['tgt_data']))
-    else:
-        return None
-
-
-class TrainValDataTransform(object):
-    """Transform the machine translation dataset.
-
-    Clip source and the target sentences to the maximum length. For the source sentence, append the
-    EOS. For the target sentence, append BOS and EOS.
+class TransformerEncoderCell(HybridBlock):
+    """Structure of the Transformer Encoder Cell.
 
     Parameters
     ----------
-    src_vocab : Vocab
-    tgt_vocab : Vocab
-    src_max_len : int
-    tgt_max_len : int
+    attention_cell : AttentionCell or str, default 'multi_head'
+        Arguments of the attention cell.
+        Can be 'multi_head', 'scaled_luong', 'scaled_dot', 'dot', 'cosine', 'normed_mlp', 'mlp'
+    units : int
+    hidden_size : int
+        number of units in the hidden layer of position-wise feed-forward networks
+    num_heads : int
+        Number of heads in multi-head attention
+    scaled : bool
+        Whether to scale the softmax input by the sqrt of the input dimension
+        in multi-head attention
+    dropout : float
+    use_residual : bool
+    output_attention: bool
+        Whether to output the attention weights
+    weight_initializer : str or Initializer
+        Initializer for the input weights matrix, used for the linear
+        transformation of the inputs.
+    bias_initializer : str or Initializer
+        Initializer for the bias vector.
+    prefix : str, default 'rnn_'
+        Prefix for name of `Block`s
+        (and name of weight if params is `None`).
+    params : Parameter or None
+        Container for weight sharing between cells.
+        Created if `None`.
     """
+    def __init__(self, attention_cell='multi_head', units=128,
+                 hidden_size=512, num_heads=4, scaled=True,
+                 dropout=0.0, use_residual=True, output_attention=False,
+                 weight_initializer=None, bias_initializer='zeros',
+                 prefix=None, params=None):
+        super(TransformerEncoderCell, self).__init__(prefix=prefix, params=params)
+        self._units = units
+        self._num_heads = num_heads
+        self._dropout = dropout
+        self._use_residual = use_residual
+        self._output_attention = output_attention
+        with self.name_scope():
+            self.dropout_layer = nn.Dropout(dropout)
+            self.attention_cell = _get_attention_cell(attention_cell,
+                                                      units=units,
+                                                      num_heads=num_heads,
+                                                      scaled=scaled,
+                                                      dropout=dropout)
+            self.proj = nn.Dense(units=units, flatten=False, use_bias=False,
+                                 weight_initializer=weight_initializer,
+                                 bias_initializer=bias_initializer,
+                                 prefix='proj_')
+            self.ffn = PositionwiseFFN(hidden_size=hidden_size, units=units,
+                                       use_residual=use_residual, dropout=dropout,
+                                       weight_initializer=weight_initializer,
+                                       bias_initializer=bias_initializer)
+            self.layer_norm = nn.LayerNorm()
 
-    def __init__(self, src_vocab, tgt_vocab, src_max_len, tgt_max_len):
-        self._src_vocab = src_vocab
-        self._tgt_vocab = tgt_vocab
-        self._src_max_len = src_max_len
-        self._tgt_max_len = tgt_max_len
+    def hybrid_forward(self, F, inputs, mask=None):  # pylint: disable=unused-argument
+        """Transformer Encoder Attention Cell.
 
-    def __call__(self, src, tgt):
-        if self._src_max_len > 0:
-            src_sentence = self._src_vocab[src.split()[:self._src_max_len]]
-        else:
-            src_sentence = self._src_vocab[src.split()]
-        if self._tgt_max_len > 0:
-            tgt_sentence = self._tgt_vocab[tgt.split()[:self._tgt_max_len]]
-        else:
-            tgt_sentence = self._tgt_vocab[tgt.split()]
-        src_sentence.append(self._src_vocab[self._src_vocab.eos_token])
-        tgt_sentence.insert(0, self._tgt_vocab[self._tgt_vocab.bos_token])
-        tgt_sentence.append(self._tgt_vocab[self._tgt_vocab.eos_token])
-        src_npy = np.array(src_sentence, dtype=np.int32)
-        tgt_npy = np.array(tgt_sentence, dtype=np.int32)
-        return src_npy, tgt_npy
+        Parameters
+        ----------
+        inputs : Symbol or NDArray
+            Input sequence. Shape (batch_size, length, C_in)
+        mask : Symbol or NDArray or None
+            Mask for inputs. Shape (batch_size, length, length)
+
+        Returns
+        -------
+        encoder_cell_outputs: list
+            Outputs of the encoder cell. Contains:
+
+            - outputs of the transformer encoder cell. Shape (batch_size, length, C_out)
+            - additional_outputs of all the transformer encoder cell
+        """
+        outputs, attention_weights =\
+            self.attention_cell(inputs, inputs, inputs, mask)
+        outputs = self.proj(outputs)
+        outputs = self.dropout_layer(outputs)
+        if self._use_residual:
+            outputs = outputs + inputs
+        outputs = self.layer_norm(outputs)
+        outputs = self.ffn(outputs)
+        additional_outputs = []
+        if self._output_attention:
+            additional_outputs.append(attention_weights)
+        return outputs, additional_outputs
 
 
-def process_dataset(dataset, src_vocab, tgt_vocab, src_max_len=-1, tgt_max_len=-1):
-    start = time.time()
-    dataset_processed = dataset.transform(TrainValDataTransform(src_vocab, tgt_vocab,
-                                                                src_max_len,
-                                                                tgt_max_len), lazy=False)
-    end = time.time()
-    print('Processing Time spent: {}'.format(end - start))
-    return dataset_processed
-
-
-def load_translation_data(dataset, src_lang='en', tgt_lang='vi'):
-    """Load translation dataset
+class TransformerDecoderCell(HybridBlock):
+    """Structure of the Transformer Decoder Cell.
 
     Parameters
     ----------
-    dataset : str
-    src_lang : str, default 'en'
-    tgt_lang : str, default 'vi'
-
-    Returns
-    -------
-
+    attention_cell : AttentionCell or str, default 'multi_head'
+        Arguments of the attention cell.
+        Can be 'multi_head', 'scaled_luong', 'scaled_dot', 'dot', 'cosine', 'normed_mlp', 'mlp'
+    units : int
+    hidden_size : int
+        number of units in the hidden layer of position-wise feed-forward networks
+    num_heads : int
+        Number of heads in multi-head attention
+    scaled : bool
+        Whether to scale the softmax input by the sqrt of the input dimension
+        in multi-head attention
+    dropout : float
+    use_residual : bool
+    output_attention: bool
+        Whether to output the attention weights
+    weight_initializer : str or Initializer
+        Initializer for the input weights matrix, used for the linear
+        transformation of the inputs.
+    bias_initializer : str or Initializer
+        Initializer for the bias vector.
+    prefix : str, default 'rnn_'
+        Prefix for name of `Block`s
+        (and name of weight if params is `None`).
+    params : Parameter or None
+        Container for weight sharing between cells.
+        Created if `None`.
     """
-    if dataset == 'IWSLT2015':
-        common_prefix = 'IWSLT2015_{}_{}_{}_{}'.format(src_lang, tgt_lang,
-                                                       args.src_max_len, args.tgt_max_len)
-        data_train = IWSLT2015('train', src_lang=src_lang, tgt_lang=tgt_lang)
-        data_val = IWSLT2015('val', src_lang=src_lang, tgt_lang=tgt_lang)
-        data_test = IWSLT2015('test', src_lang=src_lang, tgt_lang=tgt_lang)
-    elif dataset == 'WMT2016BPE':
-        common_prefix = 'WMT2016BPE_{}_{}_{}_{}'.format(src_lang, tgt_lang,
-                                                        args.src_max_len, args.tgt_max_len)
-        data_train = WMT2016BPE('train', src_lang=src_lang, tgt_lang=tgt_lang)
-        data_val = WMT2016BPE('newstest2013', src_lang=src_lang, tgt_lang=tgt_lang)
-        data_test = WMT2016BPE('newstest2014', src_lang=src_lang, tgt_lang=tgt_lang)
-    else:
-        raise NotImplementedError
-    src_vocab, tgt_vocab = data_train.src_vocab, data_train.tgt_vocab
-    data_train_processed = load_cached_dataset(common_prefix + '_train')
-    if not data_train_processed:
-        data_train_processed = process_dataset(data_train, src_vocab, tgt_vocab,
-                                               args.src_max_len, args.tgt_max_len)
-        cache_dataset(data_train_processed, common_prefix + '_train')
-    data_val_processed = load_cached_dataset(common_prefix + '_val')
-    if not data_val_processed:
-        data_val_processed = process_dataset(data_val, src_vocab, tgt_vocab)
-        cache_dataset(data_val_processed, common_prefix + '_val')
-    data_test_processed = load_cached_dataset(common_prefix + '_test')
-    if not data_test_processed:
-        data_test_processed = process_dataset(data_test, src_vocab, tgt_vocab)
-        cache_dataset(data_test_processed, common_prefix + '_test')
-    fetch_tgt_sentence = lambda src, tgt: tgt.split()
-    val_tgt_sentences = list(data_val.transform(fetch_tgt_sentence))
-    test_tgt_sentences = list(data_test.transform(fetch_tgt_sentence))
-    return data_train_processed, data_val_processed, data_test_processed, \
-           val_tgt_sentences, test_tgt_sentences, src_vocab, tgt_vocab
+    def __init__(self, attention_cell='multi_head', units=128,
+                 hidden_size=512, num_heads=4, scaled=True,
+                 dropout=0.0, use_residual=True, output_attention=False,
+                 weight_initializer=None, bias_initializer='zeros',
+                 prefix=None, params=None):
+        super(TransformerDecoderCell, self).__init__(prefix=prefix, params=params)
+        self._units = units
+        self._num_heads = num_heads
+        self._dropout = dropout
+        self._use_residual = use_residual
+        self._output_attention = output_attention
+        self._scaled = scaled
+        with self.name_scope():
+            self.dropout_layer = nn.Dropout(dropout)
+            self.attention_cell_in = _get_attention_cell(attention_cell,
+                                                         units=units,
+                                                         num_heads=num_heads,
+                                                         scaled=scaled,
+                                                         dropout=dropout)
+            self.attention_cell_inter = _get_attention_cell(attention_cell,
+                                                            units=units,
+                                                            num_heads=num_heads,
+                                                            scaled=scaled,
+                                                            dropout=dropout)
+            self.proj_in = nn.Dense(units=units, flatten=False,
+                                    use_bias=False,
+                                    weight_initializer=weight_initializer,
+                                    bias_initializer=bias_initializer,
+                                    prefix='proj_in_')
+            self.proj_inter = nn.Dense(units=units, flatten=False,
+                                       use_bias=False,
+                                       weight_initializer=weight_initializer,
+                                       bias_initializer=bias_initializer,
+                                       prefix='proj_inter_')
+            self.ffn = PositionwiseFFN(hidden_size=hidden_size,
+                                       units=units,
+                                       use_residual=use_residual,
+                                       dropout=dropout,
+                                       weight_initializer=weight_initializer,
+                                       bias_initializer=bias_initializer)
+
+            self.layer_norm_in = nn.LayerNorm()
+            self.layer_norm_inter = nn.LayerNorm()
+
+    def hybrid_forward(self, F, inputs, mem_value, mask=None, mem_mask=None):  #pylint: disable=unused-argument
+        """Transformer Decoder Attention Cell.
+
+        Parameters
+        ----------
+        inputs : Symbol or NDArray
+            Input sequence. Shape (batch_size, length, C_in)
+        mem_value : Symbol or NDArrays
+            Memory value, i.e. output of the encoder. Shape (batch_size, mem_length, C_in)
+        mask : Symbol or NDArray or None
+            Mask for inputs. Shape (batch_size, length, length)
+        mem_mask : Symbol or NDArray or None
+            Mask for mem_value. Shape (batch_size, length, mem_length)
+
+        Returns
+        -------
+        decoder_cell_outputs: list
+            Outputs of the decoder cell. Contains:
+
+            - outputs of the transformer decoder cell. Shape (batch_size, length, C_out)
+            - additional_outputs of all the transformer decoder cell
+        """
+        outputs, attention_in_outputs =\
+            self.attention_cell_in(inputs, inputs, inputs, mask)
+        outputs = self.proj_in(outputs)
+        outputs = self.dropout_layer(outputs)
+        if self._use_residual:
+            outputs = outputs + inputs
+        outputs = self.layer_norm_in(outputs)
+        inputs = outputs
+        outputs, attention_inter_outputs = \
+            self.attention_cell_inter(inputs, mem_value, mem_value, mem_mask)
+        outputs = self.proj_inter(outputs)
+        outputs = self.dropout_layer(outputs)
+        if self._use_residual:
+            outputs = outputs + inputs
+        outputs = self.layer_norm_inter(outputs)
+        outputs = self.ffn(outputs)
+        additional_outputs = []
+        if self._output_attention:
+            additional_outputs.append(attention_in_outputs)
+            additional_outputs.append(attention_inter_outputs)
+        return outputs, additional_outputs
 
 
-def get_data_lengths(dataset):
-    return list(dataset.transform(lambda srg, tgt: (len(srg), len(tgt))))
-
-
-data_train, data_val, data_test, val_tgt_sentences, test_tgt_sentences, src_vocab, tgt_vocab \
-    = load_translation_data(dataset=args.dataset, src_lang=args.src_lang, tgt_lang=args.tgt_lang)
-data_train_lengths = get_data_lengths(data_train)
-data_val_lengths = get_data_lengths(data_val)
-data_test_lengths = get_data_lengths(data_test)
-
-with io.open(os.path.join(args.save_dir, 'val_gt.txt'), 'w') as of:
-    for ele in val_tgt_sentences:
-        of.write(' '.join(ele) + '\n')
-
-with io.open(os.path.join(args.save_dir, 'test_gt.txt'), 'w') as of:
-    for ele in test_tgt_sentences:
-        of.write(' '.join(ele) + '\n')
-
-data_train = data_train.transform(lambda src, tgt: (src, tgt, len(src), len(tgt)), lazy=False)
-data_val = SimpleDataset([(ele[0], ele[1], len(ele[0]), len(ele[1]), i)
-                          for i, ele in enumerate(data_val)])
-data_test = SimpleDataset([(ele[0], ele[1], len(ele[0]), len(ele[1]), i)
-                           for i, ele in enumerate(data_test)])
-
-ctx = [mx.cpu()] if args.gpus is None or args.gpus == '' else \
-    [mx.gpu(int(x)) for x in args.gpus.split(',')]
-
-assert args.batch_size % len(ctx) == 0, \
-    'Total batch size must be multiple of the number of devices'
-
-if args.src_max_len <= 0 or args.tgt_max_len <= 0:
-    max_len = np.max(
-        [np.max(data_train_lengths, axis=0), np.max(data_val_lengths, axis=0),
-         np.max(data_test_lengths, axis=0)],
-        axis=0)
-if args.src_max_len > 0:
-    src_max_len = args.src_max_len
-else:
-    src_max_len = max_len[0]
-if args.tgt_max_len > 0:
-    tgt_max_len = args.tgt_max_len
-else:
-    tgt_max_len = max_len[1]
-encoder, decoder = get_transformer_encoder_decoder(units=args.num_units,
-                                                   hidden_size=args.hidden_size,
-                                                   dropout=args.dropout,
-                                                   num_layers=args.num_layers,
-                                                   num_heads=args.num_heads,
-                                                   max_src_length=max(src_max_len, 500),
-                                                   max_tgt_length=max(tgt_max_len, 500),
-                                                   scaled=args.scaled)
-model = NMTModel(src_vocab=src_vocab, tgt_vocab=tgt_vocab, encoder=encoder, decoder=decoder,
-                 share_embed=True, embed_size=args.num_units, tie_weights=True,
-                 embed_initializer=None, prefix='parallel_transformer_')
-model.initialize(init=mx.init.Xavier(magnitude=args.magnitude), ctx=ctx)
-static_alloc = True
-model.hybridize(static_alloc=static_alloc)
-logging.info(model)
-
-translator = BeamSearchTranslator(model=model, beam_size=args.beam_size,
-                                  scorer=BeamSearchScorer(alpha=args.lp_alpha,
-                                                          K=args.lp_k),
-                                  max_length=200)
-logging.info('Use beam_size={}, alpha={}, K={}'.format(args.beam_size, args.lp_alpha, args.lp_k))
-
-label_smoothing = LabelSmoothing(epsilon=args.epsilon, units=len(tgt_vocab))
-label_smoothing.hybridize(static_alloc=static_alloc)
-
-loss_function = SoftmaxCEMaskedLoss(sparse_label=False)
-loss_function.hybridize(static_alloc=static_alloc)
-
-test_loss_function = SoftmaxCEMaskedLoss()
-test_loss_function.hybridize(static_alloc=static_alloc)
-
-
-def evaluate(data_loader, context=ctx[0]):
-    """Evaluate given the data loader
+class TransformerEncoder(HybridBlock, Seq2SeqEncoder):
+    """Structure of the Transformer Encoder.
 
     Parameters
     ----------
-    data_loader : DataLoader
-
-    Returns
-    -------
-    avg_loss : float
-        Average loss
-    real_translation_out : list of list of str
-        The translation output
+    attention_cell : AttentionCell or str, default 'multi_head'
+        Arguments of the attention cell.
+        Can be 'multi_head', 'scaled_luong', 'scaled_dot', 'dot', 'cosine', 'normed_mlp', 'mlp'
+    num_layers : int
+    units : int
+    hidden_size : int
+        number of units in the hidden layer of position-wise feed-forward networks
+    max_length : int
+        Maximum length of the input sequence
+    num_heads : int
+        Number of heads in multi-head attention
+    scaled : bool
+        Whether to scale the softmax input by the sqrt of the input dimension
+        in multi-head attention
+    dropout : float
+    use_residual : bool
+    output_attention: bool
+        Whether to output the attention weights
+    weight_initializer : str or Initializer
+        Initializer for the input weights matrix, used for the linear
+        transformation of the inputs.
+    bias_initializer : str or Initializer
+        Initializer for the bias vector.
+    prefix : str, default 'rnn_'
+        Prefix for name of `Block`s
+        (and name of weight if params is `None`).
+    params : Parameter or None
+        Container for weight sharing between cells.
+        Created if `None`.
     """
-    translation_out = []
-    all_inst_ids = []
-    avg_loss_denom = 0
-    avg_loss = 0.0
-    for _, (src_seq, tgt_seq, src_valid_length, tgt_valid_length, inst_ids) \
-            in enumerate(data_loader):
-        src_seq = src_seq.as_in_context(context)
-        tgt_seq = tgt_seq.as_in_context(context)
-        src_valid_length = src_valid_length.as_in_context(context)
-        tgt_valid_length = tgt_valid_length.as_in_context(context)
-        # Calculating Loss
-        out, _ = model(src_seq, tgt_seq[:, :-1], src_valid_length, tgt_valid_length - 1)
-        loss = test_loss_function(out, tgt_seq[:, 1:], tgt_valid_length - 1).mean().asscalar()
-        all_inst_ids.extend(inst_ids.asnumpy().astype(np.int32).tolist())
-        avg_loss += loss * (tgt_seq.shape[1] - 1)
-        avg_loss_denom += (tgt_seq.shape[1] - 1)
-        # Translate
-        samples, _, sample_valid_length = \
-            translator.translate(src_seq=src_seq, src_valid_length=src_valid_length)
-        max_score_sample = samples[:, 0, :].asnumpy()
-        sample_valid_length = sample_valid_length[:, 0].asnumpy()
-        for i in range(max_score_sample.shape[0]):
-            translation_out.append(
-                [tgt_vocab.idx_to_token[ele] for ele in
-                 max_score_sample[i][1:(sample_valid_length[i] - 1)]])
-    avg_loss = avg_loss / avg_loss_denom
-    real_translation_out = [None for _ in range(len(all_inst_ids))]
-    for ind, sentence in zip(all_inst_ids, translation_out):
-        real_translation_out[ind] = sentence
-    return avg_loss, real_translation_out
+    def __init__(self, attention_cell='multi_head', num_layers=2,
+                 units=512, hidden_size=2048, max_length=50,
+                 num_heads=4, scaled=True, dropout=0.0,
+                 use_residual=True, output_attention=False,
+                 weight_initializer=None, bias_initializer='zeros',
+                 prefix=None, params=None):
+        super(TransformerEncoder, self).__init__(prefix=prefix, params=params)
+        assert units % num_heads == 0,\
+            'In TransformerEncoder, The units should be divided exactly ' \
+            'by the number of heads. Received units={}, num_heads={}' \
+            .format(units, num_heads)
+        self._num_layers = num_layers
+        self._max_length = max_length
+        self._num_heads = num_heads
+        self._units = units
+        self._hidden_size = hidden_size
+        self._output_attention = output_attention
+        self._dropout = dropout
+        self._use_residual = use_residual
+        self._scaled = scaled
+        with self.name_scope():
+            self.dropout_layer = nn.Dropout(dropout)
+            self.layer_norm = nn.LayerNorm()
+            self.position_weight = self.params.get_constant('const',
+                                                            _position_encoding_init(max_length,
+                                                                                    units))
+            self.transformer_cells = nn.HybridSequential()
+            for i in range(num_layers):
+                self.transformer_cells.add(
+                    TransformerEncoderCell(
+                        units=units,
+                        hidden_size=hidden_size,
+                        num_heads=num_heads,
+                        attention_cell=attention_cell,
+                        weight_initializer=weight_initializer,
+                        bias_initializer=bias_initializer,
+                        dropout=dropout,
+                        use_residual=use_residual,
+                        scaled=scaled,
+                        output_attention=output_attention,
+                        prefix='transformer%d_' % i))
 
+    def __call__(self, inputs, states=None, valid_length=None):
+        """Encoder the inputs given the states and valid sequence length.
 
-def write_sentences(sentences, file_path):
-    with io.open(file_path, 'w', encoding='utf-8') as of:
-        for sent in sentences:
-            of.write(' '.join(sent) + '\n')
+        Parameters
+        ----------
+        inputs : NDArray
+            Input sequence. Shape (batch_size, length, C_in)
+        states : list of NDArrays or None
+            Initial states. The list of initial states and masks
+        valid_length : NDArray or None
+            Valid lengths of each sequence. This is usually used when part of sequence has
+            been padded. Shape (batch_size,)
 
+        Returns
+        -------
+        encoder_outputs: list
+            Outputs of the encoder. Contains:
 
-def train():
-    """Training function."""
-    trainer = gluon.Trainer(model.collect_params(), args.optimizer,
-                            {'learning_rate': args.lr, 'beta2': 0.98, 'epsilon': 1e-9})
+            - outputs of the transformer encoder. Shape (batch_size, length, C_out)
+            - additional_outputs of all the transformer encoder
+        """
+        return super(TransformerEncoder, self).__call__(inputs, states, valid_length)
 
-    train_batchify_fn = btf.Tuple(btf.Pad(), btf.Pad(), btf.Stack(), btf.Stack())
-    test_batchify_fn = btf.Tuple(btf.Pad(), btf.Pad(), btf.Stack(), btf.Stack(), btf.Stack())
-    target_val_lengths = list(map(lambda x: x[-1], data_val_lengths))
-    target_test_lengths = list(map(lambda x: x[-1], data_test_lengths))
-    train_batch_sampler = FixedBucketSampler(lengths=data_train_lengths,
-                                             batch_size=args.batch_size,
-                                             num_buckets=args.num_buckets,
-                                             ratio=args.bucket_ratio,
-                                             shuffle=True,
-                                             use_average_length=True)
-    logging.info('Train Batch Sampler:\n{}'.format(train_batch_sampler.stats()))
-    train_data_loader = DataLoader(data_train,
-                                   batch_sampler=train_batch_sampler,
-                                   batchify_fn=train_batchify_fn,
-                                   num_workers=8)
+    def forward(self, inputs, states=None, valid_length=None, steps=None): # pylint: disable=arguments-differ
+        """
 
-    val_batch_sampler = FixedBucketSampler(lengths=target_val_lengths,
-                                           batch_size=args.test_batch_size,
-                                           num_buckets=args.num_buckets,
-                                           ratio=args.bucket_ratio,
-                                           shuffle=False,
-                                           use_average_length=True)
-    logging.info('Valid Batch Sampler:\n{}'.format(val_batch_sampler.stats()))
-    val_data_loader = DataLoader(data_val,
-                                 batch_sampler=val_batch_sampler,
-                                 batchify_fn=test_batchify_fn,
-                                 num_workers=8)
-    test_batch_sampler = FixedBucketSampler(lengths=target_test_lengths,
-                                            batch_size=args.test_batch_size,
-                                            num_buckets=args.num_buckets,
-                                            ratio=args.bucket_ratio,
-                                            shuffle=False,
-                                            use_average_length=True)
-    logging.info('Test Batch Sampler:\n{}'.format(test_batch_sampler.stats()))
-    test_data_loader = DataLoader(data_test,
-                                  batch_sampler=test_batch_sampler,
-                                  batchify_fn=test_batchify_fn,
-                                  num_workers=8)
-    best_valid_bleu = 0.0
-    step_num = 0
-    warmup_steps = args.warmup_steps
-    grad_interval = args.num_accumulated
-    model.collect_params().setattr('grad_req', 'add')
-    average_start = (len(train_data_loader) // grad_interval) * (args.epochs - args.average_start)
-    average_param_dict = None
-    model.collect_params().zero_grad()
-    for epoch_id in range(args.epochs):
-        log_avg_loss = 0
-        log_wc = 0
-        loss_denom = 0
-        step_loss = 0
-        log_start_time = time.time()
-        for batch_id, (src_seq, tgt_seq, src_valid_length, tgt_valid_length) \
-                in enumerate(train_data_loader):
-            if batch_id % grad_interval == 0:
-                step_num += 1
-                new_lr = args.lr / math.sqrt(args.num_units) \
-                         * min(1. / math.sqrt(step_num), step_num * warmup_steps ** (-1.5))
-                trainer.set_learning_rate(new_lr)
-            # logging.info(src_seq.context) Context suddenly becomes GPU.
-            src_wc = src_valid_length.sum().asscalar()
-            tgt_wc = tgt_valid_length.sum().asscalar()
-            loss_denom += tgt_wc - tgt_valid_length.shape[0]
-            if src_seq.shape[0] > len(ctx):
-                src_seq_list, tgt_seq_list, src_valid_length_list, tgt_valid_length_list \
-                    = [gluon.utils.split_and_load(seq, ctx, batch_axis=0, even_split=False)
-                       for seq in [src_seq, tgt_seq, src_valid_length, tgt_valid_length]]
+        Parameters
+        ----------
+        inputs : NDArray, Shape(batch_size, length, C_in)
+        states : list of NDArray
+        valid_length : NDArray
+        steps : NDArray
+            Stores value [0, 1, ..., length].
+            It is used for lookup in positional encoding matrix
+
+        Returns
+        -------
+        outputs : NDArray
+            The output of the encoder. Shape is (batch_size, length, C_out)
+        additional_outputs : list
+            Either be an empty list or contains the attention weights in this step.
+            The attention weights will have shape (batch_size, length, length) or
+            (batch_size, num_heads, length, length)
+
+        """
+        length = inputs.shape[1]
+        if valid_length is not None:
+            mask = mx.nd.broadcast_lesser(
+                mx.nd.arange(length, ctx=valid_length.context).reshape((1, -1)),
+                valid_length.reshape((-1, 1)))
+            mask = mx.nd.broadcast_axes(mx.nd.expand_dims(mask, axis=1), axis=1, size=length)
+            if states is None:
+                states = [mask]
             else:
-                src_seq_list = [src_seq.as_in_context(ctx[0])]
-                tgt_seq_list = [tgt_seq.as_in_context(ctx[0])]
-                src_valid_length_list = [src_valid_length.as_in_context(ctx[0])]
-                tgt_valid_length_list = [tgt_valid_length.as_in_context(ctx[0])]
+                states.append(mask)
+        inputs = inputs * math.sqrt(inputs.shape[-1])
+        steps = mx.nd.arange(length, ctx=inputs.context)
+        if states is None:
+            states = [steps]
+        else:
+            states.append(steps)
+        if valid_length is not None:
+            step_output, additional_outputs =\
+                super(TransformerEncoder, self).forward(inputs, states, valid_length)
+        else:
+            step_output, additional_outputs =\
+                super(TransformerEncoder, self).forward(inputs, states)
+        return step_output, additional_outputs
 
-            Ls = []
-            with mx.autograd.record():
-                for src_seq, tgt_seq, src_valid_length, tgt_valid_length \
-                        in zip(src_seq_list, tgt_seq_list,
-                               src_valid_length_list, tgt_valid_length_list):
-                    out, _ = model(src_seq, tgt_seq[:, :-1],
-                                   src_valid_length, tgt_valid_length - 1)
-                    smoothed_label = label_smoothing(tgt_seq[:, 1:])
-                    ls = loss_function(out, smoothed_label, tgt_valid_length - 1).sum()
-                    Ls.append((ls * (tgt_seq.shape[1] - 1)) / args.batch_size)
-            for L in Ls:
-                L.backward()
-            if batch_id % grad_interval == grad_interval - 1 or\
-                    batch_id == len(train_data_loader) - 1:
-                if average_param_dict is None:
-                    average_param_dict = {k: v.data(ctx[0]).copy() for k, v in
-                                          model.collect_params().items()}
-                trainer.step(float(loss_denom) / args.batch_size)
-                param_dict = model.collect_params()
-                param_dict.zero_grad()
-                if step_num > average_start:
-                    alpha = max(1, 1. / (step_num - average_start))
-                    for name, average_param in average_param_dict.items():
-                        average_param[:] += alpha * (param_dict[name].data(ctx[0]) - average_param)
-            step_loss += sum([L.asscalar() for L in Ls])
-            if batch_id % grad_interval == grad_interval - 1 or\
-                    batch_id == len(train_data_loader) - 1:
-                log_avg_loss += step_loss / loss_denom * args.batch_size
-                loss_denom = 0
-                step_loss = 0
-            log_wc += src_wc + tgt_wc
-            if (batch_id + 1) % (args.log_interval * grad_interval) == 0:
-                wps = log_wc / (time.time() - log_start_time)
-                logging.info('[Epoch {} Batch {}/{}] loss={:.4f}, ppl={:.4f}, '
-                             'throughput={:.2f}K wps, wc={:.2f}K'
-                             .format(epoch_id, batch_id + 1, len(train_data_loader),
-                                     log_avg_loss / args.log_interval,
-                                     np.exp(log_avg_loss / args.log_interval),
-                                     wps / 1000, log_wc / 1000))
-                log_start_time = time.time()
-                log_avg_loss = 0
-                log_wc = 0
-        mx.nd.waitall()
-        valid_loss, valid_translation_out = evaluate(val_data_loader, ctx[0])
-        valid_bleu_score, _, _, _, _ = compute_bleu([val_tgt_sentences], valid_translation_out,
-                                                    bpe=True, split_compound_word=True)
-        logging.info('[Epoch {}] valid Loss={:.4f}, valid ppl={:.4f}, valid bleu={:.2f}'
-                     .format(epoch_id, valid_loss, np.exp(valid_loss), valid_bleu_score * 100))
-        test_loss, test_translation_out = evaluate(test_data_loader, ctx[0])
-        test_bleu_score, _, _, _, _ = compute_bleu([test_tgt_sentences], test_translation_out,
-                                                   bpe=True, split_compound_word=True)
-        logging.info('[Epoch {}] test Loss={:.4f}, test ppl={:.4f}, test bleu={:.2f}'
-                     .format(epoch_id, test_loss, np.exp(test_loss), test_bleu_score * 100))
-        write_sentences(valid_translation_out,
-                        os.path.join(args.save_dir, 'epoch{:d}_valid_out.txt').format(epoch_id))
-        write_sentences(test_translation_out,
-                        os.path.join(args.save_dir, 'epoch{:d}_test_out.txt').format(epoch_id))
-        if valid_bleu_score > best_valid_bleu:
-            best_valid_bleu = valid_bleu_score
-            save_path = os.path.join(args.save_dir, 'valid_best.params')
-            logging.info('Save best parameters to {}'.format(save_path))
-            model.save_params(save_path)
-        save_path = os.path.join(args.save_dir, 'epoch{:d}.params'.format(epoch_id))
-        model.save_params(save_path)
-    save_path = os.path.join(args.save_dir, 'average.params')
-    mx.nd.save(save_path, average_param_dict)
-    if args.average_checkpoint:
-        for j in range(args.num_averages):
-            params = mx.nd.load(os.path.join(args.save_dir,
-                                             'epoch{:d}.params'.format(args.epochs - j - 1)))
-            alpha = 1. / (j + 1)
-            for k, v in model._collect_params_with_prefix().items():
-                for c in ctx:
-                    v.data(c)[:] += alpha * (params[k].as_in_context(c) - v.data(c))
-    elif args.average_start > 0:
-        for k, v in model.collect_params.items():
-            for c in ctx:
-                v.data(c)[:] = average_param_dict[k].as_in_context(c)
-    else:
-        model.load_params(os.path.join(args.save_dir, 'valid_best.params'), ctx)
-    valid_loss, valid_translation_out = evaluate(val_data_loader, ctx[0])
-    valid_bleu_score, _, _, _, _ = compute_bleu([val_tgt_sentences], valid_translation_out,
-                                                bpe=True, split_compound_word=True)
-    logging.info('Best model valid Loss={:.4f}, valid ppl={:.4f}, valid bleu={:.2f}'
-                 .format(valid_loss, np.exp(valid_loss), valid_bleu_score * 100))
-    test_loss, test_translation_out = evaluate(test_data_loader, ctx[0])
-    test_bleu_score, _, _, _, _ = compute_bleu([test_tgt_sentences], test_translation_out,
-                                               bpe=True, split_compound_word=True)
-    logging.info('Best model test Loss={:.4f}, test ppl={:.4f}, test bleu={:.2f}'
-                 .format(test_loss, np.exp(test_loss), test_bleu_score * 100))
-    write_sentences(valid_translation_out,
-                    os.path.join(args.save_dir, 'best_valid_out.txt'))
-    write_sentences(test_translation_out,
-                    os.path.join(args.save_dir, 'best_test_out.txt'))
+    def hybrid_forward(self, F, inputs, states=None, valid_length=None, position_weight=None): # pylint: disable=arguments-differ
+        """
+
+        Parameters
+        ----------
+        inputs : NDArray or Symbol, Shape(batch_size, length, C_in)
+        states : list of NDArray or Symbol
+        valid_length : NDArray or Symbol
+        position_weight : NDArray or Symbol
+
+        Returns
+        -------
+        outputs : NDArray or Symbol
+            The output of the encoder. Shape is (batch_size, length, C_out)
+        additional_outputs : list
+            Either be an empty list or contains the attention weights in this step.
+            The attention weights will have shape (batch_size, length, length) or
+            (batch_size, num_heads, length, length)
+
+        """
+        if states is not None:
+            steps = states[-1]
+            # Positional Encoding
+            inputs = F.broadcast_add(inputs, F.expand_dims(F.Embedding(steps, position_weight,
+                                                                       self._max_length,
+                                                                       self._units), axis=0))
+        inputs = self.dropout_layer(inputs)
+        inputs = self.layer_norm(inputs)
+        outputs = inputs
+        if valid_length is not None:
+            mask = states[-2]
+        else:
+            mask = None
+        additional_outputs = []
+        for cell in self.transformer_cells:
+            outputs, attention_weights = cell(inputs, mask)
+            inputs = outputs
+            if self._output_attention:
+                additional_outputs.append(attention_weights)
+        if valid_length is not None:
+            outputs = F.SequenceMask(outputs, sequence_length=valid_length,
+                                     use_sequence_length=True, axis=1)
+        return outputs, additional_outputs
 
 
-if __name__ == '__main__':
-    train()
+class TransformerDecoder(HybridBlock, Seq2SeqDecoder):
+    """Structure of the Transformer Decoder.
+
+    Parameters
+    ----------
+    attention_cell : AttentionCell or str, default 'multi_head'
+        Arguments of the attention cell.
+        Can be 'multi_head', 'scaled_luong', 'scaled_dot', 'dot', 'cosine', 'normed_mlp', 'mlp'
+    num_layers : int
+    units : int
+    hidden_size : int
+        number of units in the hidden layer of position-wise feed-forward networks
+    max_length : int
+        Maximum length of the input sequence. This is used for constructing position encoding
+    num_heads : int
+        Number of heads in multi-head attention
+    scaled : bool
+        Whether to scale the softmax input by the sqrt of the input dimension
+        in multi-head attention
+    dropout : float
+    use_residual : bool
+    output_attention: bool
+        Whether to output the attention weights
+    weight_initializer : str or Initializer
+        Initializer for the input weights matrix, used for the linear
+        transformation of the inputs.
+    bias_initializer : str or Initializer
+        Initializer for the bias vector.
+    prefix : str, default 'rnn_'
+        Prefix for name of `Block`s
+        (and name of weight if params is `None`).
+    params : Parameter or None
+        Container for weight sharing between cells.
+        Created if `None`.
+    """
+    def __init__(self, attention_cell='multi_head', num_layers=2,
+                 units=128, hidden_size=2048, max_length=50,
+                 num_heads=4, scaled=True, dropout=0.0,
+                 use_residual=True, output_attention=False,
+                 weight_initializer=None, bias_initializer='zeros',
+                 prefix=None, params=None):
+        super(TransformerDecoder, self).__init__(prefix=prefix, params=params)
+        assert units % num_heads == 0, 'In TransformerDecoder, the units should be divided ' \
+                                       'exactly by the number of heads. Received units={}, ' \
+                                       'num_heads={}'.format(units, num_heads)
+        self._num_layers = num_layers
+        self._units = units
+        self._hidden_size = hidden_size
+        self._num_states = num_heads
+        self._max_length = max_length
+        self._dropout = dropout
+        self._use_residual = use_residual
+        self._output_attention = output_attention
+        self._scaled = scaled
+        with self.name_scope():
+            self.dropout_layer = nn.Dropout(dropout)
+            self.layer_norm = nn.LayerNorm()
+            self.position_weight = self.params.get_constant('const',
+                                                            _position_encoding_init(max_length,
+                                                                                    units))
+            self.transformer_cells = nn.HybridSequential()
+            for i in range(num_layers):
+                self.transformer_cells.add(
+                    TransformerDecoderCell(
+                        units=units,
+                        hidden_size=hidden_size,
+                        num_heads=num_heads,
+                        attention_cell=attention_cell,
+                        weight_initializer=weight_initializer,
+                        bias_initializer=bias_initializer,
+                        dropout=dropout,
+                        scaled=scaled,
+                        use_residual=use_residual,
+                        output_attention=output_attention,
+                        prefix='transformer%d_' % i))
+
+    def init_state_from_encoder(self, encoder_outputs, encoder_valid_length=None):
+        """Initialize the state from the encoder outputs.
+
+        Parameters
+        ----------
+        encoder_outputs : list
+        encoder_valid_length : NDArray or None
+
+        Returns
+        -------
+        decoder_states : list
+            The decoder states, includes:
+
+            - mem_value : NDArray
+            - mem_masks : NDArray, optional
+        """
+        mem_value = encoder_outputs
+        decoder_states = [mem_value]
+        mem_length = mem_value.shape[1]
+        if encoder_valid_length is not None:
+            mem_masks = mx.nd.broadcast_lesser(
+                mx.nd.arange(mem_length, ctx=encoder_valid_length.context).reshape((1, -1)),
+                encoder_valid_length.reshape((-1, 1)))
+            decoder_states.append(mem_masks)
+        self._encoder_valid_length = encoder_valid_length
+        return decoder_states
+
+    def decode_seq(self, inputs, states, valid_length=None):
+        """Decode the decoder inputs. This function is only used for training.
+
+        Parameters
+        ----------
+        inputs : NDArray, Shape (batch_size, length, C_in)
+        states : list of NDArrays or None
+            Initial states. The list of decoder states
+        valid_length : NDArray or None
+            Valid lengths of each sequence. This is usually used when part of sequence has
+            been padded. Shape (batch_size,)
+
+        Returns
+        -------
+        output : NDArray, Shape (batch_size, length, C_out)
+        states : list
+            The decoder states, includes:
+
+            - mem_value : NDArray
+            - mem_masks : NDArray, optional
+        additional_outputs : list of list
+            Either be an empty list or contains the attention weights in this step.
+            The attention weights will have shape (batch_size, length, mem_length) or
+            (batch_size, num_heads, length, mem_length)
+        """
+        batch_size = inputs.shape[0]
+        length = inputs.shape[1]
+        length_array = mx.nd.arange(length, ctx=inputs.context)
+        mask = mx.nd.broadcast_lesser_equal(
+            length_array.reshape((1, -1)),
+            length_array.reshape((-1, 1)))
+        if valid_length is not None:
+            batch_mask = mx.nd.broadcast_lesser(
+                mx.nd.arange(length, ctx=valid_length.context).reshape((1, -1)),
+                valid_length.reshape((-1, 1)))
+            mask = mx.nd.broadcast_mul(mx.nd.expand_dims(batch_mask, -1),
+                                       mx.nd.expand_dims(mask, 0))
+        else:
+            mask = mx.nd.broadcast_axes(mx.nd.expand_dims(mask, axis=0), axis=0, size=batch_size)
+        states = [None] + states
+        output, states, additional_outputs = self.forward(inputs, states, mask)
+        states = states[1:]
+        if valid_length is not None:
+            output = mx.nd.SequenceMask(output,
+                                        sequence_length=valid_length,
+                                        use_sequence_length=True,
+                                        axis=1)
+        return output, states, additional_outputs
+
+    def __call__(self, step_input, states):
+        """One-step-ahead decoding of the Transformer decoder.
+
+        Parameters
+        ----------
+        step_input : NDArray
+        states : list of NDArray
+
+        Returns
+        -------
+        step_output : NDArray
+            The output of the decoder.
+            In the train mode, Shape is (batch_size, length, C_out)
+            In the test mode, Shape is (batch_size, C_out)
+        new_states: list
+            Includes
+            - last_embeds : NDArray or None
+                It is only given during testing
+            - mem_value : NDArray
+            - mem_masks : NDArray, optional
+
+        step_additional_outputs : list of list
+            Either be an empty list or contains the attention weights in this step.
+            The attention weights will have shape (batch_size, length, mem_length) or
+            (batch_size, num_heads, length, mem_length)
+        """
+        return super(TransformerDecoder, self).__call__(step_input, states)
+
+    def forward(self, step_input, states, mask=None):  #pylint: disable=arguments-differ, missing-docstring
+        input_shape = step_input.shape
+        mem_mask = None
+        # If it is in testing, transform input tensor to a tensor with shape NTC
+        # Otherwise remove the None in states.
+        if len(input_shape) == 2:
+            if self._encoder_valid_length is not None:
+                has_last_embeds = len(states) == 3
+            else:
+                has_last_embeds = len(states) == 2
+            if has_last_embeds:
+                last_embeds = states[0]
+                step_input = mx.nd.concat(last_embeds,
+                                          mx.nd.expand_dims(step_input, axis=1),
+                                          dim=1)
+                states = states[1:]
+            else:
+                step_input = mx.nd.expand_dims(step_input, axis=1)
+        elif states[0] is None:
+            states = states[1:]
+        has_mem_mask = (len(states) == 2)
+        if has_mem_mask:
+            _, mem_mask = states
+            augmented_mem_mask = mx.nd.expand_dims(mem_mask, axis=1)\
+                .broadcast_axes(axis=1, size=step_input.shape[1])
+            states[-1] = augmented_mem_mask
+        if mask is None:
+            length_array = mx.nd.arange(step_input.shape[1], ctx=step_input.context)
+            mask = mx.nd.broadcast_lesser_equal(
+                length_array.reshape((1, -1)),
+                length_array.reshape((-1, 1)))
+            mask = mx.nd.broadcast_axes(mx.nd.expand_dims(mask, axis=0),
+                                        axis=0, size=step_input.shape[0])
+        steps = mx.nd.arange(step_input.shape[1], ctx=step_input.context)
+        states.append(steps)
+        step_output, step_additional_outputs = \
+            super(TransformerDecoder, self).forward(step_input * math.sqrt(step_input.shape[-1]),  #pylint: disable=too-many-function-args
+                                                    states, mask)
+        states = states[:-1]
+        if has_mem_mask:
+            states[-1] = mem_mask
+        new_states = [step_input] + states
+        # If it is in testing, only output the last one
+        if len(input_shape) == 2:
+            step_output = step_output[:, -1, :]
+        return step_output, new_states, step_additional_outputs
+
+    def hybrid_forward(self, F, step_input, states, mask=None, position_weight=None):  #pylint: disable=arguments-differ
+        """
+
+        Parameters
+        ----------
+        step_input : NDArray or Symbol, Shape (batch_size, length, C_in)
+        states : list of NDArray or Symbol
+        mask : NDArray or Symbol
+        position_weight : NDArray or Symbol
+
+        Returns
+        -------
+        step_output : NDArray or Symbol
+            The output of the decoder. Shape is (batch_size, length, C_out)
+        step_additional_outputs : list
+            Either be an empty list or contains the attention weights in this step.
+            The attention weights will have shape (batch_size, length, mem_length) or
+            (batch_size, num_heads, length, mem_length)
+
+        """
+        has_mem_mask = (len(states) == 3)
+        if has_mem_mask:
+            mem_value, mem_mask, steps = states
+        else:
+            mem_value, steps = states
+            mem_mask = None
+        # Positional Encoding
+        step_input = F.broadcast_add(step_input,
+                                     F.expand_dims(F.Embedding(steps,
+                                                               position_weight,
+                                                               self._max_length,
+                                                               self._units),
+                                                   axis=0))
+        step_input = self.dropout_layer(step_input)
+        step_input = self.layer_norm(step_input)
+        inputs = step_input
+        outputs = inputs
+        step_additional_outputs = []
+        attention_weights_l = []
+        for cell in self.transformer_cells:
+            outputs, attention_weights = cell(inputs, mem_value, mask, mem_mask)
+            if self._output_attention:
+                attention_weights_l.append(attention_weights)
+            inputs = outputs
+        if self._output_attention:
+            step_additional_outputs.extend(attention_weights_l)
+        return outputs, step_additional_outputs
+
+
+def get_transformer_encoder_decoder(num_layers=2,
+                                    num_heads=5, scaled=True,
+                                    units=512, hidden_size=2048, dropout=0.0, use_residual=True,
+                                    max_src_length=50, max_tgt_length=50,
+                                    weight_initializer=None, bias_initializer='zeros',
+                                    prefix='transformer_', params=None):
+    """Build a pair of Parallel GNMT encoder/decoder
+
+    Parameters
+    ----------
+    num_layers : int
+    num_heads : int
+    scaled : bool
+    units : int
+    hidden_size : int
+    dropout : float
+    use_residual : bool
+    max_src_length : int
+    max_tgt_length : int
+    weight_initializer : mx.init.Initializer or None
+    bias_initializer : mx.init.Initializer or None
+    prefix : str, default 'transformer_'
+        Prefix for name of `Block`s.
+    params : Parameter or None
+        Container for weight sharing between layers.
+        Created if `None`.
+
+    Returns
+    -------
+    encoder : TransformerEncoder
+    decoder :TransformerDecoder
+    """
+    encoder = TransformerEncoder(num_layers=num_layers,
+                                 num_heads=num_heads,
+                                 max_length=max_src_length,
+                                 units=units,
+                                 hidden_size=hidden_size,
+                                 dropout=dropout,
+                                 scaled=scaled,
+                                 use_residual=use_residual,
+                                 weight_initializer=weight_initializer,
+                                 bias_initializer=bias_initializer,
+                                 prefix=prefix + 'enc_', params=params)
+    decoder = TransformerDecoder(num_layers=num_layers,
+                                 num_heads=num_heads,
+                                 max_length=max_tgt_length,
+                                 units=units,
+                                 hidden_size=hidden_size,
+                                 dropout=dropout,
+                                 scaled=scaled,
+                                 use_residual=use_residual,
+                                 weight_initializer=weight_initializer,
+                                 bias_initializer=bias_initializer,
+                                 prefix=prefix + 'dec_', params=params)
+    return encoder, decoder
